% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{age_abbrv_to_name}
\alias{age_abbrv_to_name}
\title{Convert age-class abbreviations to full names}
\usage{
age_abbrv_to_name(abbrv)
}
\arguments{
\item{abbrv}{A character vector of age-class abbreviations.}
}
\value{
A character vector of the same length as \code{abbrv}, containing the
corresponding age-class names. Unknown abbreviations return \code{NA}.
}
\description{
Translates standardised age-class abbreviations into their corresponding
descriptive category names.
}
\examples{
age_abbrv_to_name(c("Ad", "J", "Ni"))

}
