% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azores.bkw-package.R
\docType{package}
\name{azores.bkw-package}
\alias{azores.bkw}
\alias{azores.bkw-package}
\title{azores.bkw: R Compendium Accompanying Peres dos Santos et al. (2026)}
\description{
An R Compendium providing data, code and documentation needed to replicate the results of the paper Peres dos Santos, R., Sears, R., Magno, R. and Castilho, R. xxxxx xxxxxxxxxxxxxxxxxxxxxxxxxxx xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx. JOURNAL XX, DDD–DDD (YYYY). \doi{xx.xxxx/xxxxxx-xxx-xxxxx-x}.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Rui Peres dos Santos \email{rpsantos@pattern.institute} (\href{https://orcid.org/0000-0002-5979-0700}{ORCID}) [data contributor]
  \item Rita Castilho \email{rcastil@ualg.pt} (\href{https://orcid.org/0000-0003-0727-3688}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item CCMAR [copyright holder, funder]
}

}
\keyword{internal}
