% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{behaviour_abbrv_to_name}
\alias{behaviour_abbrv_to_name}
\title{Convert behaviour abbreviations to full names}
\usage{
behaviour_abbrv_to_name(abbrv)
}
\arguments{
\item{abbrv}{A character vector of behaviour abbreviations.}
}
\value{
A character vector of the same length as \code{abbrv}, containing the
corresponding behaviour names. Unknown abbreviations return \code{NA}.
}
\description{
Translates standardised behaviour abbreviations into their corresponding
descriptive behaviour names.
}
\examples{
behaviour_abbrv_to_name(c("FO", "R", "TF", "Ni"))

}
