% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_date_time.R
\name{combine_date_time}
\alias{combine_date_time}
\title{Combine date and time into a UTC datetime}
\usage{
combine_date_time(date, time, tz = "UTC")
}
\arguments{
\item{date}{A vector coercible to \code{Date}.}

\item{time}{A vector coercible to POSIXct/POSIXlt, from which hour, minute,
and second components will be extracted.}

\item{tz}{Time zone for the resulting datetime. Defaults to \code{"UTC"}.}
}
\value{
A POSIXct vector.
}
\description{
Constructs a POSIXct datetime by combining the calendar components of a
\code{Date} object with the clock components of a time-like object.
}
\details{
This helper is intended for datasets where dates and times are stored in
separate columns.
}
\keyword{internal}
