% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{extract_hmlmeso}
\alias{extract_hmlmeso}
\title{Extract hmlmeso values at sf locations}
\usage{
extract_hmlmeso(sf, dir)
}
\arguments{
\item{sf}{An \code{sf} object containing a \code{date} column.}

\item{dir}{Directory containing the hmlmeso NetCDF file.}
}
\value{
An \code{sf} object with the same columns as the input,
plus an additional column named \code{"hmlmeso"} placed before
the geometry column.
}
\description{
\code{\link[=extract_hmlmeso]{extract_hmlmeso()}} extracts highly migrant lower mesopelagic micronekton (hmlmeso) values
for each geometry in an \code{sf} object based on the corresponding
\code{date} column.

Extraction is performed by grouping records by date and matching
each group to the appropriate raster layer.
}
