% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{extract_lmeso}
\alias{extract_lmeso}
\title{Extract lmeso values at sf locations}
\usage{
extract_lmeso(sf, dir)
}
\arguments{
\item{sf}{An \code{sf} object containing a \code{date} column.}

\item{dir}{Directory containing the hmlmeso NetCDF file.}
}
\value{
An \code{sf} object with the same columns as the input,
plus an additional column named \code{"lmeso"} placed before
the geometry column.
}
\description{
\code{\link[=extract_lmeso]{extract_lmeso()}} extracts lower mesopelagic micronekton (lmeso) values
for each geometry in an \code{sf} object based on the corresponding
\code{date} column.

Extraction is performed by grouping records by date and matching
each group to the appropriate raster layer.
}
