% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{hmlmeso_dates}
\alias{hmlmeso_dates}
\title{Index ocean wet mass content of highly migrant lower mesopelagic micronekton "hmlmeso" raster layers by date}
\usage{
hmlmeso_dates()
}
\value{
A tibble with columns:
\describe{
\item{date}{Daily study dates.}
\item{day}{Row index of the date sequence.}
\item{file}{Name of the NetCDF file containing hmlmeso data.}
\item{i}{Layer index within the NetCDF file corresponding to each date.}
}
}
\description{
\code{\link[=hmlmeso_dates]{hmlmeso_dates()}} creates a lookup table linking each study date to
the corresponding layer index in the hmlmeso NetCDF file.
}
