% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{mlmeso_dates}
\alias{mlmeso_dates}
\title{Index ocean wet mass content of migrant lower mesopelagic micronekton "mlmeso" raster layers by date}
\usage{
mlmeso_dates()
}
\value{
A tibble with columns:
\describe{
\item{date}{Daily study dates.}
\item{day}{Row index of the date sequence.}
\item{file}{Name of the NetCDF file containing mlmeso data.}
\item{i}{Layer index within the NetCDF file corresponding to each date.}
}
}
\description{
\code{\link[=mlmeso_dates]{mlmeso_dates()}} creates a lookup table linking each study date to
the corresponding layer index in the mlmeso NetCDF file.
}
