% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{reaction_abbrv_to_name}
\alias{reaction_abbrv_to_name}
\title{Convert reaction abbreviations to full names}
\usage{
reaction_abbrv_to_name(abbrv)
}
\arguments{
\item{abbrv}{A character vector of reaction abbreviations.}
}
\value{
A character vector of the same length as \code{abbrv}, containing the
corresponding reaction descriptions. Unknown abbreviations return \code{NA}.
}
\description{
Translates abbreviations describing cetacean reactions to a vessel into
descriptive reaction categories.
}
\details{
Non-informative reaction codes (\code{"Ni"}, \code{"NI"}) are recoded to \code{"other"}.
}
\examples{
reaction_abbrv_to_name(c("A", "I", "E", "NI"))

}
