% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{sst_dates}
\alias{sst_dates}
\title{Index sea surface temperature "sst" raster layers by date}
\usage{
sst_dates()
}
\value{
A tibble with columns:
\describe{
\item{date}{Daily study dates.}
\item{day}{Row index of the date sequence.}
\item{file}{Name of the NetCDF file containing SST data.}
\item{i}{Layer index within the NetCDF file corresponding to each date.}
}
}
\description{
\code{\link[=sst_dates]{sst_dates()}} creates a lookup table linking each study date to
the corresponding layer index in the SST NetCDF file.
}
