% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkw.extract.R
\name{study_dates}
\alias{study_dates}
\title{Generate study date sequence}
\usage{
study_dates(from = "2012-01-01", to = "2018-12-31")
}
\arguments{
\item{from}{A character or \code{Date} object indicating the start date.
Defaults to \code{"2012-01-01"}.}

\item{to}{A character or \code{Date} object indicating the end date.
Defaults to \code{"2018-12-31"}.}
}
\value{
A vector of class \code{Date} with daily resolution.
}
\description{
\code{\link[=study_dates]{study_dates()}} generates a daily sequence of dates between the
specified start and end dates. This sequence is used to index
environmental raster time series.
}
